+++ components/frontend/src/header_footer/Menubar.js
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import { Button, Container, Form, Header, Icon, Image, Input, Menu, Message, Modal } from 'semantic-ui-react';
 import { DateInput } from 'semantic-ui-calendar-react';
+import md5 from 'md5';
 import './Menubar.css';
 
 function Login(props) {
@@ -24,7 +25,9 @@ function Login(props) {
 
 function Logout(props) {
   return (
-    <Button secondary onClick={props.logout}><Icon name='user' />Logout {props.user}</Button>
+    <Button secondary onClick={props.logout}>
+      <Image avatar src={`https://www.gravatar.com/avatar/${md5(props.email || '')}?d=identicon`} />Logout {props.user}
+    </Button>
   )
 }
 
@@ -60,7 +63,7 @@ export function Menubar(props) {
             />
           </Menu.Item>
           <Menu.Item>
-            {(props.user !== null) ? <Logout user={props.user} logout={props.logout} /> : <Login login={props.login} error={props.login_error} />}
+            {(props.user !== null) ? <Logout email={props.email} user={props.user} logout={props.logout} /> : <Login login={props.login} error={props.login_error} />}
           </Menu.Item>
         </Menu.Menu>
       </Container>